fix: cli use MesBuilder settings.

+++ src/main.rs
@@ -1,8 +1,11 @@
-use std::{path::{PathBuf}};
+use std::{path::{PathBuf}, io::Read};
 
-use clap::{Parser, Subcommand};
+use clap::{Parser, Subcommand, builder};
 use config::{Config, File};
-use mes::mes::{MeSConfig, get_default_config};
+//use mes::mes::{get_default_config};
+use mes::mes::builder::{MeSBuilder};
+use question::{Question, Answer};
+use std::io;
 
 #[derive(Debug, Parser)]
 #[clap(name = "mes", author, about, version)]
@@ -49,15 +52,15 @@ enum Commands {
 fn main() {
     let cli = Cli::parse();
     //コンフィグの初期化
-    let mes_conf: MeSConfig;
+    let mes_conf: MeSBuilder;
     if std::path::Path::exists(std::path::Path::new(&cli.conf)){
         mes_conf = Config::builder()
             .add_source(File::with_name(&cli.conf))
             .build()
             .unwrap()
-            .try_deserialize::<MeSConfig>().unwrap();
+            .try_deserialize::<MeSBuilder>().unwrap();
     }else{
-        mes_conf = get_default_config();
+        mes_conf = mes::mes::builder::new();
     }
 
     //サブコマンドの解析   
@@ -99,22 +102,33 @@ fn do_chat(path: PathBuf){
     //let content = std::fs::read_to_string(path).expect("could not read file");
     println!("chatはまだ実装されていません.");
 }
-fn do_count(path: PathBuf, conf: MeSConfig){
+fn do_count(path: PathBuf, conf: MeSBuilder){
     let content = std::fs::read_to_string(path).expect("could not read file");
-    let json = mes::countDialogueWordToJsonWithConf(&content, conf);
+    let json = mes::mes::countDialogueWordToJsonWithConf(content, conf);
     println!("{}", json);
 }
 
 fn do_config_create(path: String){
     //let filepath = PathBuf::from(&path);
-    let defConf = get_default_config();
+    let def_conf = mes::mes::builder::new();
     //let json = r#"{"name": "sample","counter": {"ignore_char": ["a","b","c"]}}"#;
-    let json = serde_json::to_string(&defConf).unwrap();
-    //TODO: すでにファイルがある場合は上書きするか確認をする
+    let json = serde_json::to_string_pretty(&def_conf).unwrap();
+    //すでにファイルがある場合は上書きするか確認をする
+    let filepath = std::path::Path::new(&path);
+    if std::path::Path::exists(filepath) {
+        let answer = Question::new("すでにファイルが存在します。上書きしますか？")
+            .confirm();
+        
+        if answer == Answer::NO {
+            return //早期リターン
+        }
+    }
     std::fs::write(path, &json).expect("cannot write config");
+ 
+
 }
 
-fn do_config_show(path: String, mesconf: MeSConfig){
+fn do_config_show(path: String, mesconf: MeSBuilder){
     
     println!("{}", serde_json::to_string(&mesconf).unwrap());
